Refactor CLI convert to new module
